Add files via upload
Medical research university X is undergoing a deep research on patients with certain conditions.  Universityhas an internal AI team. Due to confidentiality the patient’s details and the conditions are masked by  the client by providing different datasets to the AI team for developing a AIML model which can predict thecondition of the patient depending on the received test results. Demonstrating the ability to fetch, process and leverage data to generate useful predictions  by trainingSupervised Learning algorithms.